pyproject.toml
[tool.poetry]
name = "printclaw"
version = "0.1.0"
description = "Local-first AI printer troubleshooting agent with a web GUI"
authors = ["Printclaw Team <support@printclaw.local>"]
readme = "README.md"
packages = [{ include = "printclaw" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
typer = "^0.12.5"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
structlog = "^24.4.0"
pyyaml = "^6.0.2"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
ruff = "^0.6.9"
black = "^24.8.0"

[tool.poetry.scripts]
printclaw = "printclaw.cli:app"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
